aaaa/requirements.txt
numpy
opencv-python
sewar
pillow

